Modified requests
Script allows user to log in if there credentials are not available in a pkl file. If they're not available it allows user to log in and saves credentials for the next time they run he application. Makes call the Calendar API and returns the next 10 events in the users calendar with the title and start time